Changes divider to border for CSS-only handling
The divider was part of the sign in area links. This commit removes the
divider from the HTML and makes it a border instead of a pipe "|". This
leads to a slightly taller divider, but it is also more of an even
height with the text than the pipe was.